add richard derivative :)

+++ src/ex5/solutions/integral.py
@@ -5,8 +5,8 @@ from src.ex5.utils.gauss_quadrature import get_gauss_legendre_quadrature
 class Integral(Solution):
   def __init__(self, c1, c2, c3, c4, **kwargs):
     super().__init__(c1, c2, c3, c4, **kwargs)
-    self.a = int(input('a: '))
-    self.b = int(input('b: '))
+    self.a = float(input('a: '))
+    self.b = float(input('b: '))
     self.chosen_method = int(input('Escolha um método - 1 para Gauss-Legendre; 2 para Quadratura Polinomial: '))
     if self.chosen_method not in [1, 2]:
       raise RuntimeError('Método inválido.')
